// src/models/packages/PackageModel.ts
export interface PackageModel {
  packageName?: string | undefined;
  endDate?: string | undefined;
  companyLogo?: string | undefined;
  phoneNumber?: any;
  id?: number;
  companyName?: string | undefined;
  duration?: string;
  expDate?: string;
  price?: number;
  priceDisplayValue?: string;
  amount?: number;
}